requirements.txt
tinkoff-investments==0.2.0b109
numpy==1.22.4
pandas==2.2.3
python-dotenv==1.0.1